math changes

+++ src/headers/celestial.hpp
@@ -12,8 +12,8 @@ class Celestial {
 	virtual std::string getName() = 0;
 	virtual float getMass() = 0;
 	virtual vector getPos() = 0;
-	virtual void updateVel(std::vector<Celestial *>, unsigned int) = 0;
-	virtual void updatePos(std::vector<Celestial *>, unsigned int) = 0;
+	virtual void updateVel(std::vector<Celestial *>, double) = 0;
+	virtual void updatePos(std::vector<Celestial *>, double) = 0;
 	virtual void draw(Shader &) = 0;
 };
